// backend-service/apps/deployment-scenario/src/deployment-scenario/api/di-token.ts
export enum DeploymentScenarioDIToken {
  DeploymentScenarioService = 'DEPLOYMENTSCENARIO_SERVICE',
  DeploymentScenarioRepository = 'DEPLOYMENTSCENARIO_REPOSITORY',

  AttackerService = 'ATTACKER_SERVICE',
  AttackerRepository = 'ATTACKER_REPOSITORY',

  AppliedCountermeasureRepository = 'APPLIED_COUNTERMEASURE_REPOSITORY',
  AppliedCountermeasureService = 'APPLIED_COUNTERMEASURE_SERVICE',

  AssetCveService = 'ASSETCVE_SERVICE',
  AssetCveRepository = 'ASSETCVE_REPOSITORY',

  DeploymentScenarioPublisher = 'DEPLOYMENTSCENARIO_PUBLISHER',

  RiskCoreService = 'RISK_CORE_SERVICE',

  SystemProfileClientProxy = 'SYSTEM_PROFILE_CLIENT_PROXY',
  AssetManagementClientProxy = 'ASSET_MANAGEMENT_CLIENT_PROXY',
  CategoryClientProxy = 'CATEGORY_CLIENT_PROXY',
  SupplementAptRiskClientProxy = 'SUPPLEMENT_APT_RISK_CLIENT_PROXY',

  ContextService = 'CONTEXT_SERVICE',
}